NP-3442 Remove unused styles, and rename to mainTheme
