+++ src/pages/registration/resource_type_tab/components/JournalField.tsx
@@ -1,12 +1,11 @@
 import { Field, FieldProps, useFormikContext } from 'formik';
 import { useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import { Chip, ThemeProvider, Typography } from '@mui/material';
+import { Chip, Typography } from '@mui/material';
 import { Autocomplete } from '@mui/material';
 import { AutocompleteTextField } from '../../../../components/AutocompleteTextField';
 import { EmphasizeSubstring } from '../../../../components/EmphasizeSubstring';
 import { StyledFlexColumn } from '../../../../components/styled/Wrappers';
-import { lightTheme } from '../../../../themes/lightTheme';
 import { Journal, PublicationChannelType } from '../../../../types/registration.types';
 import { useFetch } from '../../../../utils/hooks/useFetch';
 import { PublicationChannelApiPath } from '../../../../api/apiPaths';
@@ -68,92 +67,90 @@ export const JournalField = () => {
   );
 
   return (
-    <ThemeProvider theme={lightTheme}>
-      <Field name={ResourceFieldNames.PublicationContextId}>
-        {({ field, meta }: FieldProps<string>) => (
-          <Autocomplete
-            multiple
-            id={journalFieldTestId}
-            data-testid={journalFieldTestId}
-            aria-labelledby={`${journalFieldTestId}-label`}
-            popupIcon={null}
-            options={debouncedQuery && query === debouncedQuery && !isLoadingJournalOptions ? journalOptions ?? [] : []}
-            filterOptions={(options) => options}
-            inputValue={query}
-            onInputChange={(_, newInputValue, reason) => {
-              if (reason !== 'reset') {
-                setQuery(newInputValue);
-              }
-              if (reason === 'input' && !newInputValue && reference?.publicationContext.title) {
-                setFieldValue(contextTypeBaseFieldName, { type: PublicationChannelType.UnconfirmedSeries });
-              }
-            }}
-            onBlur={() => setFieldTouched(field.name, true, false)}
-            blurOnSelect
-            disableClearable={!query}
-            value={reference?.publicationContext.id && journal ? [journal] : []}
-            onChange={(_, inputValue, reason) => {
-              if (reason === 'selectOption') {
-                setFieldValue(contextTypeBaseFieldName, {
-                  type: PublicationChannelType.Journal,
-                  id: inputValue.pop()?.id,
-                });
-              } else if (reason === 'removeOption') {
-                setFieldValue(contextTypeBaseFieldName, { type: PublicationChannelType.UnconfirmedJournal });
-              }
-              setQuery('');
-            }}
-            loading={isLoadingJournalOptions || isLoadingJournal}
-            getOptionLabel={(option) => option.name}
-            renderOption={(props, option, state) => (
-              <li {...props}>
-                <StyledFlexColumn>
-                  <Typography variant="subtitle1">
-                    <EmphasizeSubstring
-                      text={getPublicationChannelString(option.name, option.onlineIssn, option.printIssn)}
-                      emphasized={state.inputValue}
-                    />
+    <Field name={ResourceFieldNames.PublicationContextId}>
+      {({ field, meta }: FieldProps<string>) => (
+        <Autocomplete
+          multiple
+          id={journalFieldTestId}
+          data-testid={journalFieldTestId}
+          aria-labelledby={`${journalFieldTestId}-label`}
+          popupIcon={null}
+          options={debouncedQuery && query === debouncedQuery && !isLoadingJournalOptions ? journalOptions ?? [] : []}
+          filterOptions={(options) => options}
+          inputValue={query}
+          onInputChange={(_, newInputValue, reason) => {
+            if (reason !== 'reset') {
+              setQuery(newInputValue);
+            }
+            if (reason === 'input' && !newInputValue && reference?.publicationContext.title) {
+              setFieldValue(contextTypeBaseFieldName, { type: PublicationChannelType.UnconfirmedSeries });
+            }
+          }}
+          onBlur={() => setFieldTouched(field.name, true, false)}
+          blurOnSelect
+          disableClearable={!query}
+          value={reference?.publicationContext.id && journal ? [journal] : []}
+          onChange={(_, inputValue, reason) => {
+            if (reason === 'selectOption') {
+              setFieldValue(contextTypeBaseFieldName, {
+                type: PublicationChannelType.Journal,
+                id: inputValue.pop()?.id,
+              });
+            } else if (reason === 'removeOption') {
+              setFieldValue(contextTypeBaseFieldName, { type: PublicationChannelType.UnconfirmedJournal });
+            }
+            setQuery('');
+          }}
+          loading={isLoadingJournalOptions || isLoadingJournal}
+          getOptionLabel={(option) => option.name}
+          renderOption={(props, option, state) => (
+            <li {...props}>
+              <StyledFlexColumn>
+                <Typography variant="subtitle1">
+                  <EmphasizeSubstring
+                    text={getPublicationChannelString(option.name, option.onlineIssn, option.printIssn)}
+                    emphasized={state.inputValue}
+                  />
+                </Typography>
+                {option.level && (
+                  <Typography variant="body2" color="textSecondary">
+                    {t('resource_type.level')}: {option.level}
                   </Typography>
-                  {option.level && (
+                )}
+              </StyledFlexColumn>
+            </li>
+          )}
+          renderTags={(value, getTagProps) =>
+            value.map((option, index) => (
+              <Chip
+                {...getTagProps({ index })}
+                data-testid={dataTestId.registrationWizard.resourceType.journalChip}
+                label={
+                  <>
+                    <Typography variant="subtitle1">
+                      {getPublicationChannelString(option.name, option.onlineIssn, option.printIssn)}
+                    </Typography>
                     <Typography variant="body2" color="textSecondary">
                       {t('resource_type.level')}: {option.level}
                     </Typography>
-                  )}
-                </StyledFlexColumn>
-              </li>
-            )}
-            renderTags={(value, getTagProps) =>
-              value.map((option, index) => (
-                <Chip
-                  {...getTagProps({ index })}
-                  data-testid={dataTestId.registrationWizard.resourceType.journalChip}
-                  label={
-                    <>
-                      <Typography variant="subtitle1">
-                        {getPublicationChannelString(option.name, option.onlineIssn, option.printIssn)}
-                      </Typography>
-                      <Typography variant="body2" color="textSecondary">
-                        {t('resource_type.level')}: {option.level}
-                      </Typography>
-                    </>
-                  }
-                />
-              ))
-            }
-            renderInput={(params) => (
-              <AutocompleteTextField
-                {...params}
-                required
-                label={t('resource_type.journal')}
-                isLoading={isLoadingJournalOptions || isLoadingJournal}
-                placeholder={!reference?.publicationContext.id ? t('resource_type.search_for_journal') : ''}
-                showSearchIcon={!reference?.publicationContext.id}
-                errorMessage={meta.touched && !!meta.error ? meta.error : ''}
+                  </>
+                }
               />
-            )}
-          />
-        )}
-      </Field>
-    </ThemeProvider>
+            ))
+          }
+          renderInput={(params) => (
+            <AutocompleteTextField
+              {...params}
+              required
+              label={t('resource_type.journal')}
+              isLoading={isLoadingJournalOptions || isLoadingJournal}
+              placeholder={!reference?.publicationContext.id ? t('resource_type.search_for_journal') : ''}
+              showSearchIcon={!reference?.publicationContext.id}
+              errorMessage={meta.touched && !!meta.error ? meta.error : ''}
+            />
+          )}
+        />
+      )}
+    </Field>
   );
 };
