+++ src/pages/registration/resource_type_tab/components/PublisherField.tsx
@@ -1,11 +1,10 @@
 import { Field, FieldProps, useFormikContext } from 'formik';
 import { useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import { Chip, ThemeProvider, Typography } from '@mui/material';
+import { Chip, Typography } from '@mui/material';
 import { Autocomplete } from '@mui/material';
 import { AutocompleteTextField } from '../../../../components/AutocompleteTextField';
 import { EmphasizeSubstring } from '../../../../components/EmphasizeSubstring';
-import { lightTheme } from '../../../../themes/lightTheme';
 import { PublicationChannelType, Publisher, Registration } from '../../../../types/registration.types';
 import { useFetch } from '../../../../utils/hooks/useFetch';
 import { PublicationChannelApiPath } from '../../../../api/apiPaths';
@@ -54,93 +53,91 @@ export const PublisherField = () => {
   );
 
   return (
-    <ThemeProvider theme={lightTheme}>
-      <Field name={ResourceFieldNames.PublicationContextPublisherId}>
-        {({ field, meta }: FieldProps<string>) => (
-          <Autocomplete
-            multiple
-            id={publisherFieldTestId}
-            data-testid={publisherFieldTestId}
-            aria-labelledby={`${publisherFieldTestId}-label`}
-            popupIcon={null}
-            options={
-              debouncedQuery && query === debouncedQuery && !isLoadingPublisherOptions ? publisherOptions ?? [] : []
+    <Field name={ResourceFieldNames.PublicationContextPublisherId}>
+      {({ field, meta }: FieldProps<string>) => (
+        <Autocomplete
+          multiple
+          id={publisherFieldTestId}
+          data-testid={publisherFieldTestId}
+          aria-labelledby={`${publisherFieldTestId}-label`}
+          popupIcon={null}
+          options={
+            debouncedQuery && query === debouncedQuery && !isLoadingPublisherOptions ? publisherOptions ?? [] : []
+          }
+          filterOptions={(options) => options}
+          inputValue={query}
+          onInputChange={(_, newInputValue, reason) => {
+            if (reason !== 'reset') {
+              setQuery(newInputValue);
             }
-            filterOptions={(options) => options}
-            inputValue={query}
-            onInputChange={(_, newInputValue, reason) => {
-              if (reason !== 'reset') {
-                setQuery(newInputValue);
-              }
-              if (reason === 'input' && !newInputValue && publisher?.name) {
-                setFieldValue(ResourceFieldNames.PublicationContextPublisher, {
-                  type: PublicationChannelType.UnconfirmedPublisher,
-                });
-              }
-            }}
-            onBlur={() => setFieldTouched(field.name, true, false)}
-            blurOnSelect
-            disableClearable={!query}
-            value={publisher?.id && fetchedPublisher ? [fetchedPublisher] : []}
-            onChange={(_, inputValue, reason) => {
-              if (reason === 'selectOption') {
-                setFieldValue(ResourceFieldNames.PublicationContextPublisher, {
-                  type: PublicationChannelType.Publisher,
-                  id: inputValue.pop()?.id,
-                });
-              } else if (reason === 'removeOption') {
-                setFieldValue(ResourceFieldNames.PublicationContextPublisher, {
-                  type: PublicationChannelType.UnconfirmedPublisher,
-                });
-              }
-              setQuery('');
-            }}
-            loading={isLoadingPublisherOptions || isLoadingPublisher}
-            getOptionLabel={(option) => option.name}
-            renderOption={(props, option, state) => (
-              <li {...props}>
-                <StyledFlexColumn>
-                  <Typography variant="subtitle1">
-                    <EmphasizeSubstring text={option.name} emphasized={state.inputValue} />
+            if (reason === 'input' && !newInputValue && publisher?.name) {
+              setFieldValue(ResourceFieldNames.PublicationContextPublisher, {
+                type: PublicationChannelType.UnconfirmedPublisher,
+              });
+            }
+          }}
+          onBlur={() => setFieldTouched(field.name, true, false)}
+          blurOnSelect
+          disableClearable={!query}
+          value={publisher?.id && fetchedPublisher ? [fetchedPublisher] : []}
+          onChange={(_, inputValue, reason) => {
+            if (reason === 'selectOption') {
+              setFieldValue(ResourceFieldNames.PublicationContextPublisher, {
+                type: PublicationChannelType.Publisher,
+                id: inputValue.pop()?.id,
+              });
+            } else if (reason === 'removeOption') {
+              setFieldValue(ResourceFieldNames.PublicationContextPublisher, {
+                type: PublicationChannelType.UnconfirmedPublisher,
+              });
+            }
+            setQuery('');
+          }}
+          loading={isLoadingPublisherOptions || isLoadingPublisher}
+          getOptionLabel={(option) => option.name}
+          renderOption={(props, option, state) => (
+            <li {...props}>
+              <StyledFlexColumn>
+                <Typography variant="subtitle1">
+                  <EmphasizeSubstring text={option.name} emphasized={state.inputValue} />
+                </Typography>
+                {option.level && (
+                  <Typography variant="body2" color="textSecondary">
+                    {t('resource_type.level')}: {option.level}
                   </Typography>
-                  {option.level && (
+                )}
+              </StyledFlexColumn>
+            </li>
+          )}
+          renderTags={(value, getTagProps) =>
+            value.map((option, index) => (
+              <Chip
+                {...getTagProps({ index })}
+                data-testid={dataTestId.registrationWizard.resourceType.publisherChip}
+                label={
+                  <>
+                    <Typography variant="subtitle1">{option.name}</Typography>
                     <Typography variant="body2" color="textSecondary">
                       {t('resource_type.level')}: {option.level}
                     </Typography>
-                  )}
-                </StyledFlexColumn>
-              </li>
-            )}
-            renderTags={(value, getTagProps) =>
-              value.map((option, index) => (
-                <Chip
-                  {...getTagProps({ index })}
-                  data-testid={dataTestId.registrationWizard.resourceType.publisherChip}
-                  label={
-                    <>
-                      <Typography variant="subtitle1">{option.name}</Typography>
-                      <Typography variant="body2" color="textSecondary">
-                        {t('resource_type.level')}: {option.level}
-                      </Typography>
-                    </>
-                  }
-                />
-              ))
-            }
-            renderInput={(params) => (
-              <AutocompleteTextField
-                {...params}
-                required
-                label={t('common:publisher')}
-                isLoading={isLoadingPublisherOptions || isLoadingPublisher}
-                placeholder={!publisher?.id ? t('resource_type.search_for_publisher') : ''}
-                showSearchIcon={!publisher?.id}
-                errorMessage={meta.touched && !!meta.error ? meta.error : ''}
+                  </>
+                }
               />
-            )}
-          />
-        )}
-      </Field>
-    </ThemeProvider>
+            ))
+          }
+          renderInput={(params) => (
+            <AutocompleteTextField
+              {...params}
+              required
+              label={t('common:publisher')}
+              isLoading={isLoadingPublisherOptions || isLoadingPublisher}
+              placeholder={!publisher?.id ? t('resource_type.search_for_publisher') : ''}
+              showSearchIcon={!publisher?.id}
+              errorMessage={meta.touched && !!meta.error ? meta.error : ''}
+            />
+          )}
+        />
+      )}
+    </Field>
   );
 };
